// src/server/messaging/whatsapp/types.ts
export interface MetaWebhookPayload {
  object: 'whatsapp_business_account'
  entry: MetaWebhookEntry[]
}

export interface MetaWebhookEntry {
  id: string
  changes: MetaWebhookChange[]
}

export interface MetaWebhookChange {
  value: MetaWebhookValue
  field: 'messages'
}